SLE-1166: Don't share the instance and dispose marker views correctly
Don't share the instance of `MarkerViewWithBottomPanel` that was always the latest view initialized.
Also dispose the view correctly as otherwise this would have triggered a exception when the instance was still around but the views were closed.
