+++ org.sonarlint.eclipse.ui/src/org/sonarlint/eclipse/ui/internal/views/issues/TaintVulnerabilitiesView.java
@@ -19,9 +19,8 @@
  */
 package org.sonarlint.eclipse.ui.internal.views.issues;
 
+import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.swt.SWT;
-import org.eclipse.swt.layout.GridData;
-import org.eclipse.swt.layout.RowLayout;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Link;
 import org.sonarlint.eclipse.core.documentation.SonarLintDocumentation;
@@ -29,24 +28,27 @@ import org.sonarlint.eclipse.ui.internal.SonarLintUiPlugin;
 import org.sonarlint.eclipse.ui.internal.util.BrowserUtils;
 
 public class TaintVulnerabilitiesView extends MarkerViewWithBottomPanel {
-
   public static final String ID = SonarLintUiPlugin.PLUGIN_ID + ".views.issues.TaintVulnerabilitiesView";
+
+  @Nullable
+  private static TaintVulnerabilitiesView instance;
 
   public TaintVulnerabilitiesView() {
     super(SonarLintUiPlugin.PLUGIN_ID + ".views.issues.taintIssueMarkerGenerator");
+    instance = this;
   }
 
   @Override
   protected void populateBottomPanel(Composite bottom) {
-    var bottomLayout = new RowLayout();
-    bottomLayout.center = true;
-    bottom.setLayout(bottomLayout);
-    var bottomLayoutData = new GridData(SWT.FILL, SWT.FILL, true, false);
-    bottom.setLayoutData(bottomLayoutData);
-
     bottomLabel = new Link(bottom, SWT.NONE);
     bottomLabel.addListener(SWT.Selection,
       e -> BrowserUtils.openExternalBrowser(SonarLintDocumentation.TAINT_VULNERABILITIES_LINK, e.display));
+  }
+
+  @Override
+  public void dispose() {
+    instance = null;
+    super.dispose();
   }
 
   @Override
@@ -55,5 +57,10 @@ public class TaintVulnerabilitiesView extends MarkerViewWithBottomPanel {
       "This view displays taint vulnerabilities detected by SonarQube (Server, Cloud). SonarQube for Eclipse does not "
         + "detect those issues locally. <a>Learn more</a>");
     bottomLabel.getParent().layout();
+  }
+
+  @Nullable
+  public static TaintVulnerabilitiesView getInstance() {
+    return instance;
   }
 }
